additional check in login route
