+++ apps/af-mvp/next.config.js
@@ -21,7 +21,7 @@ const nextSafeConfig = {
       "'self'",
       CODESETS_BASE_URL,
       'https://fonts.googleapis.com/',
-      'https://cognito-idp.eu-north-1.amazonaws.com/'
+      'https://cognito-idp.eu-north-1.amazonaws.com/',
     ],
     'default-src': "'self'",
     'font-src': ["'self'", 'https://fonts.gstatic.com/'],
@@ -73,7 +73,12 @@ const nextConfig = {
     config.plugins.push(
       new DuplicatePackageCheckerPlugin({
         exclude(instance) {
-          return ['react-is', 'tslib', '@aws-crypto/util', '@aws-crypto/sha256-js'].includes(instance.name);
+          return [
+            'react-is',
+            'tslib',
+            '@aws-crypto/util',
+            '@aws-crypto/sha256-js',
+          ].includes(instance.name);
         },
       })
     );
