+++ apps/af-mvp/src/pages/api/auth/login.route.ts
@@ -2,12 +2,22 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { loggedOutAuthMiddleware } from '@mvp/lib/backend/middleware/auth';
 import { generateCSRFToken } from '@mvp/lib/backend/secrets-and-tokens';
 import { retrieveSinunaLoginUrl } from '@mvp/lib/backend/services/sinuna/sinuna-requests';
+import { isSinunaDisabled } from '@mvp/lib/shared/sinuna-status';
 import cookie from 'cookie';
 
 export default loggedOutAuthMiddleware(async function handler(
   req: NextApiRequest,
   res: NextApiResponse
 ) {
+  if (isSinunaDisabled()) {
+    return res.redirect(
+      303,
+      `/profile?sinunaError=${encodeURIComponent(
+        'Sinuna login service has been disabled. You can use Virtual Finland login service to sign-in.'
+      )}`
+    );
+  }
+
   const sinunaState = await generateCSRFToken();
   const sinunaLoginUrl = await retrieveSinunaLoginUrl(req, sinunaState);
   res
