+++ packages/af-shared/tsconfig.json
@@ -6,7 +6,8 @@
   "compilerOptions": {
     "baseUrl": "./src",
     "paths": {
-      "@/*": ["./*"]
+      "@/*": ["./*"],
+      "@mvp/*": ["../../../apps/af-mvp/src/*"]
     }
   }
 }
